+++ lib/component/component.dart
@@ -3,7 +3,6 @@ import 'package:universally/universally.dart';
 
 export 'src/alert.dart';
 export 'src/gif.dart';
-export 'src/input_text.dart';
 export 'src/list.dart';
 export 'src/picker.dart';
 export 'src/scaffold.dart';